+++ bin/license_checker.dart
@@ -4,15 +4,12 @@ import 'package:args/command_runner.dart';
 import 'package:io/io.dart';
 import 'package:license_checker2/command_runner.dart';
 import 'package:license_checker2/src/commands/check_license.dart';
-import 'package:license_checker2/src/commands/generate_disclaimer.dart';
 import 'package:license_checker2/src/commands/utils.dart';
 
 void main(List<String> arguments) async {
   exitCode = ExitCode.success.code;
 
-  final cmd = LicenseCommandRunner()
-    ..addCommand(CheckLicenses())
-    ..addCommand(GenerateDisclaimer());
+  final cmd = LicenseCommandRunner()..addCommand(CheckLicenses());
 
   try {
     final errors = await cmd.run(arguments);