+++ lib/generate_disclaimer.dart
@@ -1,3 +1,2 @@
 export 'package:license_checker2/src/config.dart';
-export 'package:license_checker2/src/generate_disclaimer.dart';
 export 'package:license_checker2/src/package_checker.dart';